Refactor

+++ examples/fontview/print.rs
@@ -4,7 +4,7 @@ pub fn glyph(glyph: &aaf::Glyph) {
     if !glyph.pixels.is_empty() {
         for row in 0..glyph.height {
             for column in 0..glyph.width {
-                let pixel = &glyph.pixels[row * glyph.width + column];
+                let pixel = &glyph.pixels[(row * glyph.width + column) as usize];
                 let levels = [' ', '.', ':', '-', '=', '+', '*', '#', '%', '@'];
 
                 print!(
